+++ languagetool-language-modules/de/src/main/java/org/languagetool/language/AustrianGerman.java
@@ -24,6 +24,7 @@ import org.languagetool.UserConfig;
 import org.languagetool.languagemodel.LanguageModel;
 import org.languagetool.rules.Rule;
 import org.languagetool.rules.de.AustrianGermanSpellerRule;
+import org.languagetool.rules.de.GermanCompoundRule;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -41,6 +42,13 @@ public class AustrianGerman extends German {
   @Override
   public String getName() {
     return "German (Austria)";
+  }
+
+  @Override
+  public List<Rule> getRelevantRules(ResourceBundle messages, UserConfig userConfig, Language motherTongue, List<Language> altLanguages) throws IOException {
+    List<Rule> rules = new ArrayList<>(super.getRelevantRules(messages, userConfig, motherTongue, altLanguages));
+    rules.add(new GermanCompoundRule(messages));
+    return rules;
   }
 
   @Override