+++ languagetool-language-modules/de/src/test/java/org/languagetool/rules/de/GermanCompoundRuleTest.java
@@ -20,23 +20,24 @@ package org.languagetool.rules.de;
 
 import java.io.IOException;
 
-import org.junit.Before;
 import org.junit.Test;
 import org.languagetool.JLanguageTool;
 import org.languagetool.Languages;
 import org.languagetool.TestTools;
 import org.languagetool.rules.AbstractCompoundRuleTest;
 
-public class CompoundRuleTest extends AbstractCompoundRuleTest {
-
-  @Before
-  public void setUp() throws Exception {
-    lt = new JLanguageTool(Languages.getLanguageForShortCode("de-DE"));
-    rule = new CompoundRule(TestTools.getMessages("de"));
-  }
+public class GermanCompoundRuleTest extends AbstractCompoundRuleTest {
 
   @Test
   public void testRule() throws IOException {
+    lt = new JLanguageTool(Languages.getLanguageForShortCode("de-DE"));
+    rule = new GermanCompoundRule(TestTools.getMessages("de"));
+    runTests();
+    rule = new SwissCompoundRule(TestTools.getMessages("de"));
+    runTests();
+  }
+  
+  private void runTests() throws IOException {
     // correct sentences:
     check(0, "Eine tolle CD-ROM");
     check(0, "Eine tolle CD-ROM.");